+++ project/common/ExternalInterface.cpp
@@ -102,6 +102,46 @@ static value is_any_view_visible()
 }
 DEFINE_PRIM(is_any_view_visible, 0);
 
+static value get_custom_id()
+{
+	return alloc_string(getCustomID());
+}
+DEFINE_PRIM(get_custom_id, 0);
+
+static value set_should_request_interstitials_in_first_session(value shouldRequest)
+{
+	setShouldRequestInterstitialsInFirstSession(val_bool(shouldRequest));
+	return alloc_null();
+}
+DEFINE_PRIM(set_should_request_interstitials_in_first_session, 1);
+
+static value set_should_display_loading_view_for_more_apps(value shouldDisplay)
+{
+	setShouldDisplayLoadingViewForMoreApps(val_bool(shouldDisplay));
+	return alloc_null();
+}
+DEFINE_PRIM(set_should_display_loading_view_for_more_apps, 1);
+
+static value set_should_prefetch_video_content(value shouldPrefetch)
+{
+	setShouldPrefetchVideoContent(val_bool(shouldPrefetch));
+	return alloc_null();
+}
+DEFINE_PRIM(set_should_prefetch_video_content, 1);
+
+static value get_sdk_version()
+{
+	return alloc_string(getSDKVersion());
+}
+DEFINE_PRIM(get_sdk_version, 0);
+
+static value set_status_bar_behavior(value shouldHide)
+{
+	setStatusBarBehavior(val_bool(shouldHide));
+	return alloc_null();
+}
+DEFINE_PRIM(set_status_bar_behavior, 1);
+
 extern "C" void samcodeschartboost_main()
 {
 	val_int(0);
@@ -113,7 +153,7 @@ extern "C" int samcodeschartboost_register_prims()
 	return 0; 
 }
 
-extern "C" void sendChartboostEvent(const char* type, const char* location, const char* uri, int reward_coins, int error)
+extern "C" void sendChartboostEvent(const char* type, const char* location, const char* uri, int reward_coins, int error, bool status)
 {
     if(chartboostEventHandle == 0)
     {
@@ -126,7 +166,9 @@ extern "C" void sendChartboostEvent(const char* type, const char* location, cons
 	alloc_field(o, val_id("uri"), alloc_string(uri));
 	alloc_field(o, val_id("reward_coins"), alloc_int(reward_coins));
 	alloc_field(o, val_id("error"), alloc_int(error));
+	alloc_field(o, val_id("status"), alloc_bool(status));
 	val_call1(chartboostEventHandle->get(), o);
 }
 
 #endif
+